RobolectricSample/src/main/java/com/pivotallabs/GenericAdapter.java
package com.pivotallabs;

import android.view.View;
import android.view.ViewGroup;
import android.widget.BaseAdapter;

import java.util.List;

public abstract class GenericAdapter<T> extends BaseAdapter {

    private List<T> list;

    public GenericAdapter(List<T> list) {
        this.list = list;
    }

    @Override
    public int getCount() {
        return list.size();
    }

    @Override
    public T getItem(int i) {
        return list.get(i);
    }

    @Override
    public long getItemId(int i) {
        return i;
    }

    @Override
    public View getView(int position, View convertView, ViewGroup parent) {
     return getView(getItem(position), convertView, parent);
    }

    abstract public View getView(T item, View recycleView, ViewGroup parent);
}
